fix(config-plugin-import): fails file import
The command tries to import attached files two times - one time by parsing the files.json file in the data folder (which succeeds), and once more by relying on the `file` attribute in a setting in the XML.

The `file` attribute in the setting refers to a hash of the file so the lookup fails and the command crashes at this point.

The fix here simply removes the second importation logic.

The solution is not perfect as probably the best way would have been to find the hash in files.json and import only the files referenced in settings.
But it is simple and should work.