feat(plugin-install): add support for installing from zip file
It is quite weird that this option doesn't exist. There is a
plugin-download command available in Moosh and then plugin-install
downloads and installs. It seems more reasonable to separate download
and installation. Which is basically what I did here without changing
the original behaviour.

Do note that this commit is incomplete. It should work for the intended
purpose, but the code can handle edge cases and the separation of
behaviours better.